Add --max-time, --scrub, --defrag flags, wave-based memory processing, and 32-bit fixes
- Add --max-time flag for graceful time-limited dedup runs
- Add --scrub and --defrag flags for post-dedup btrfs maintenance
- Refactor default mode to wave-based collection with memory eviction
- Add scrub/defrag/max-time options to cron job config (disabled by default)
- Fix int overflow on 32-bit architectures in progress bar calculations
- Fix btrfs magic constant for 32-bit compilation
- Fix scrub polling infinite loop and defrag ETA calculation
- Fix cache save on batch mode early return and wave deferred count
- Add build constraint for lock.go (unix) and lock_stub.go (!unix)
- Add btrfs and lock tests
- Remove vendor/ from git tracking (regenerated by release.sh)

+++ progress.go
@@ -66,12 +66,12 @@ func formatCount(n int64) string {
 }
 
 // printProgressBar renders a progress bar on stderr, overwriting the current line.
-func printProgressBar(prefix string, current, total int, suffix string) {
+func printProgressBar(prefix string, current, total int64, suffix string) {
 	if quietMode || !isTTY || total <= 0 {
 		return
 	}
 	pct := current * 100 / total
-	filled := barWidth * current / total
+	filled := int(barWidth * current / total)
 	if filled > barWidth {
 		filled = barWidth
 	}
@@ -96,7 +96,7 @@ func printStatus(msg string) {
 }
 
 // formatETA returns a compact time estimate string like "1.5m", "30s", "2.3h".
-func formatETA(elapsed time.Duration, current, total int) string {
+func formatETA(elapsed time.Duration, current, total int64) string {
 	if current <= 0 || total <= 0 || current > total {
 		return ""
 	}
